requirements.txt
accelerate==0.33.0
blobfile==2.1.1
certifi==2024.7.4
charset-normalizer==3.3.2
ctransformers==0.2.27
diskcache==5.6.3
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
llama_cpp_python==0.2.87
lxml==4.9.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
psutil==6.0.0
py-cpuinfo==9.0.0
pycryptodomex==3.20.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sentencepiece==0.2.0
setuptools==72.1.0
sympy==1.13.1
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
